Fix nagios_die wrapper not properly passing last argument to nagios_exit
nagios_die() accepts many ways of specifying arguments. One of them is
passing the message only, in which case the last argument added by
Plugins.pm is on the 2nd position instead of the 3rd. One visible
effect of this bug is the inability to print custom "shortname" with
nagios_die when the exit code is not specified.